+++ RELEASE_NOTES.md
@@ -8,6 +8,7 @@ The following items are the set of security fixes that have been
 applied since Enterprise Chef 11.1.3:
 
 * [bootstrap] Set random initial password for pivotal superuser on bootstrap
+# [opscode-account] Prevent password authentication for pivotal superuser
 
 ## 11.1.3 (2014-04-09)